+++ flexmeasures/data/schemas/locations.py
